refactor: simplify token verification by extracting logic to a separate jwt-verifier module
